RESTORED ORIGINAL DATA BEFORE THE WRONG MERGE
