+++ src/components/DealsContainer.jsx
@@ -1,8 +1,7 @@
 import React from 'react'
-import HomeContainer from './HomeContainer'
 import { motion } from "framer-motion"
 import { MdChevronLeft, MdChevronRight } from 'react-icons/md'
-import RowContainer from "./RowContainer"
+import RowContainer from "./RowContainer.jsx"
 
 const DealsContainer = () => {
     return (
