+++ src/components/RowContainer.jsx
@@ -5,7 +5,7 @@ import { motion } from 'framer-motion'
 const RowContainer = ({ flag, data }) => {
     console.log(data);
     return (
-        <div className={`w-full gap-3  flex items-center my-12 bg-red-400 ${flag ? 'overflow-x-scroll' : 'overflow-x-hidden flex-wrap'}`}>
+        <div className={`w-full flex items-center gap-3 my-12 bg-gradient-to-tl from-red-500 to red-800 ${flag ? 'overflow-x-scroll' : 'overflow-x-hidden flex-wrap'}`}>
 
 
             {data && data.map(item => (< div key={item.id} className='w-300 min-w-[300px] md:w-340 md:min-w-[340px] h-auto my-12 bg-orange-400 rounded-lg p-2 shadow-md backdrop-blur-lg hover:drop-shadow-lg' >
