+++ src/components/StoresContainer.jsx
@@ -3,7 +3,6 @@ import React from 'react'
 const StoresContainer = () => {
     return (
         //Your Code for store will begin here 
-
         <section>
             <center>
                 <div>Restaurants</div>
@@ -11,8 +10,7 @@ const StoresContainer = () => {
                 <div>What are you waiting for?</div>
             </center>
         </section>
-        <div>StoresContainer</div>
-        )
+    )
 }
 
 export default StoresContainer
