+++ src/components/index.js
@@ -3,14 +3,9 @@ import { filterProps } from "framer-motion";
 export { default as Header } from "./Header";
 export {default as MainContainer} from "./MainContainer";
 export {default as CreateContainer} from "./CreateContainer";
-
 export {default as HomeContainer} from "./HomeContainer";
 export {default as StoresContainer} from "./StoresContainer";
 export {default as DealsContainer} from "./DealsContainer";
 export {default as AboutContainer} from "./AboutContainer";
 export { default as Loader } from "./Loader";
 export { default as RowContainer } from "./RowContainer";
-
-export {default as HomeContainer} from "./HomeContainer"
-
-
