+++ src/index.js
@@ -9,9 +9,6 @@ import { StateProvider } from "./context/StateProvider";
 import { initialState } from "./context/initialState";
 import reducer from "./context/reducer";
 
-import 'bootstrap/dist/css/bootstrap.min.css';
-import 'bootstrap/dist/js/bootstrap.bundle.min';
-
 ReactDOM.render(
     <Router >
         <StateProvider initialState={initialState} reducer={reducer}>
